Move the examples to the facade crate
