--- example/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "example"
-version = "0.1.0"
-authors = ["pierre <pierre.krieger1708@gmail.com>"]
-
-[dependencies]
-bigint = "4.2"
-bytes = "0.4"
-env_logger = "0.5.4"
-futures = "0.1"
-multiaddr = "0.3"
-libp2p-mplex = { path = "../mplex" }
-libp2p-identify = { path = "../identify" }
-libp2p-kad = { path = "../kad" }
-libp2p-floodsub = { path = "../floodsub" }
-libp2p-peerstore = { path = "../peerstore" }
-libp2p-ping = { path = "../ping" }
-libp2p-relay = { path = "../relay" }
-libp2p-secio = { path = "../secio" }
-libp2p-core = { path = "../core" }
-libp2p-tcp-transport = { path = "../tcp-transport" }
-libp2p-websocket = { path = "../websocket" }
-rand = "0.4"
-structopt =  "0.2"
-tokio-core = "0.1"
-tokio-io = "0.1"
-tokio-stdin = "0.1"
