+++ libp2p/examples/echo-dialer.rs
@@ -21,21 +21,20 @@
 extern crate bytes;
 extern crate env_logger;
 extern crate futures;
-extern crate libp2p_mplex as multiplex;
-extern crate libp2p_ping as ping;
-extern crate libp2p_secio as secio;
-extern crate libp2p_core as swarm;
-extern crate libp2p_tcp_transport as tcp;
+extern crate libp2p;
 extern crate tokio_core;
 extern crate tokio_io;
 
-use futures::Future;
 use futures::sync::oneshot;
+use futures::{Future, Sink, Stream};
 use std::env;
-use swarm::Transport;
-use swarm::upgrade::{self, DeniedConnectionUpgrade};
-use tcp::TcpConfig;
+use libp2p::core::Transport;
+use libp2p::core::upgrade::{self, DeniedConnectionUpgrade, SimpleProtocol};
+use libp2p::tcp::TcpConfig;
 use tokio_core::reactor::Core;
+use tokio_io::AsyncRead;
+use tokio_io::codec::BytesCodec;
+use libp2p::websocket::WsConfig;
 
 fn main() {
     env_logger::init();
@@ -43,7 +42,7 @@ fn main() {
     // Determine which address to dial.
     let target_addr = env::args()
         .nth(1)
-        .unwrap_or("/ip4/127.0.0.1/tcp/4001".to_owned());
+        .unwrap_or("/ip4/127.0.0.1/tcp/10333".to_owned());
 
     // We start by building the tokio engine that will run all the sockets.
     let mut core = Core::new().unwrap();
@@ -51,6 +50,11 @@ fn main() {
     // Now let's build the transport stack.
     // We start by creating a `TcpConfig` that indicates that we want TCP/IP.
     let transport = TcpConfig::new(core.handle())
+
+        // In addition to TCP/IP, we also want to support the Websockets protocol on top of TCP/IP.
+        // The parameter passed to `WsConfig::new()` must be an implementation of `Transport` to be
+        // used for the underlying multiaddress.
+        .or_transport(WsConfig::new(TcpConfig::new(core.handle())))
 
         // On top of TCP/IP, we will use either the plaintext protocol or the secio protocol,
         // depending on which one the remote supports.
@@ -60,8 +64,8 @@ fn main() {
             let secio = {
                 let private_key = include_bytes!("test-private-key.pk8");
                 let public_key = include_bytes!("test-public-key.der").to_vec();
-                secio::SecioConfig {
-                    key: secio::SecioKeyPair::rsa_from_pkcs8(private_key, public_key).unwrap(),
+                libp2p::secio::SecioConfig {
+                    key: libp2p::secio::SecioKeyPair::rsa_from_pkcs8(private_key, public_key).unwrap(),
                 }
             };
 
@@ -69,7 +73,7 @@ fn main() {
         })
 
         // On top of plaintext or secio, we will use the multiplex protocol.
-        .with_upgrade(multiplex::MultiplexConfig::new())
+        .with_upgrade(libp2p::mplex::MultiplexConfig::new())
         // The object returned by the call to `with_upgrade(MultiplexConfig::new())` can't be used as a
         // `Transport` because the output of the upgrade is not a stream but a controller for
         // muxing. We have to explicitly call `into_connection_reuse()` in order to turn this into
@@ -80,25 +84,41 @@ fn main() {
     // connections for us. The second parameter we pass is the connection upgrade that is accepted
     // by the listening part. We don't want to accept anything, so we pass a dummy object that
     // represents a connection that is always denied.
-    let (swarm_controller, swarm_future) = swarm::swarm(
+    let (swarm_controller, swarm_future) = libp2p::core::swarm(
         transport.clone().with_upgrade(DeniedConnectionUpgrade),
         |_socket, _client_addr| -> Result<(), _> {
             unreachable!("All incoming connections should have been denied")
         },
     );
 
+    // Building a struct that represents the protocol that we are going to use for dialing.
+    let proto = SimpleProtocol::new("/echo/1.0.0", |socket| {
+        // This closure is called whenever a stream using the "echo" protocol has been
+        // successfully negotiated. The parameter is the raw socket (implements the AsyncRead
+        // and AsyncWrite traits), and the closure must return an implementation of
+        // `IntoFuture` that can yield any type of object.
+        Ok(AsyncRead::framed(socket, BytesCodec::new()))
+    });
+
     // We now use the controller to dial to the address.
-    let (tx, rx) = oneshot::channel();
+    let (finished_tx, finished_rx) = oneshot::channel();
     swarm_controller
-        .dial_custom_handler(target_addr.parse().expect("invalid multiaddr"),
-            transport.with_upgrade(ping::Ping),
-            |(mut pinger, future), _| {
-                let ping = pinger.ping().map_err(|_| unreachable!()).inspect(|_| {
-                    println!("Received pong from the remote");
-                    let _ = tx.send(());
-                });
-                ping.select(future).map(|_| ()).map_err(|(e, _)| e)
-            })
+        .dial_custom_handler(target_addr.parse().expect("invalid multiaddr"), transport.with_upgrade(proto), |echo, _| {
+            // `echo` is what the closure used when initializing `proto` returns.
+            // Consequently, please note that the `send` method is available only because the type
+            // `length_delimited::Framed` has a `send` method.
+            println!("Sending \"hello world\" to listener");
+            echo.send("hello world".into())
+                // Then listening for one message from the remote.
+                .and_then(|echo| {
+                    echo.into_future().map_err(|(e, _)| e).map(|(n,_ )| n)
+                })
+                .and_then(|message| {
+                    println!("Received message from listener: {:?}", message.unwrap());
+                    finished_tx.send(()).unwrap();
+                    Ok(())
+                })
+        })
         // If the multiaddr protocol exists but is not supported, then we get an error containing
         // the original multiaddress.
         .expect("unsupported multiaddr");
@@ -110,9 +130,9 @@ fn main() {
     // `swarm_future` is a future that contains all the behaviour that we want, but nothing has
     // actually started yet. Because we created the `TcpConfig` with tokio, we need to run the
     // future through the tokio core.
-    core.run(
-        rx.select(swarm_future.map_err(|_| unreachable!()))
-            .map_err(|(e, _)| e)
-            .map(|_| ()),
-    ).unwrap();
+    let final_future = swarm_future
+        .select(finished_rx.map_err(|_| unreachable!()))
+        .map(|_| ())
+        .map_err(|(err, _)| err);
+    core.run(final_future).unwrap();
 }
